backend/requirements.txt
Flask==2.2.2
PyPDF2==1.26.0
pandas==1.5.3
requests==2.28.1
mysql-connector-python==8.0.32
scikit-learn==1.2.0
joblib==1.2.0
transformers==4.11.3
torch==1.9.0